+++ R/app_ui.R
@@ -3,6 +3,11 @@ shiny::addResourcePath("archeoViz", system.file("R", package="archeoViz"))
 
 ui <- shinyUI(
   fluidPage(
+    tags$head(tags$style(HTML(".tabclass 
+                              th { padding: 0 1em 1em ;}
+                              td { padding: 0 1em 1em ;}
+                              "))),
+    
     theme = shinythemes::shinytheme(getShinyOption("set.theme")),
     # Sidebar ----
     sidebarLayout(
@@ -104,8 +109,7 @@ ui <- shinyUI(
           tabPanel(.term_switcher("tab.plot3d"), # 3D plot ----
                    fluidRow(
                      column(10,
-                            plotly::plotlyOutput("plot3d",  width = "100%", height = 650),
-                            uiOutput("id.table"), 
+                            plotly::plotlyOutput("plot3d",  width = "100%", height = 650)
                      ),
                      column(2,
                             br(),
@@ -119,16 +123,21 @@ ui <- shinyUI(
                             sliderInput("point.size", .term_switcher("point.size"), width="100%", sep = "",
                                         min=1, max=5, value=2, step=1),
                             uiOutput("ratio3D"), 
-                            downloadButton("download.3d.plot", .term_switcher("download"))
+                            downloadButton("download.3d.plot", .term_switcher("export"))
                      )  # end column
-                   )  # end fluid row
+                   ),  # end fluid row
+                   fluidRow(column(12,
+                     p(.term_switcher("click.on.point")),
+                     uiOutput("plot3d.selection.tab"),
+                     br()
+                   )) # end colums, end fluidrow
           ),      #end tabPanel
 
           tabPanel(.term_switcher("tab.map"),  # map ----
                    fluidRow(
                      column(10,
                             uiOutput("sliderMap"),
-                            plotly::plotlyOutput("map", width = "100%", height = 500),
+                            plotly::plotlyOutput("map", width = "100%", height = 500)
                      ),
                      column(2,
                             br(),
@@ -139,8 +148,14 @@ ui <- shinyUI(
                             sliderInput("map.point.size", .term_switcher("point.size"),
                                         width="100%", sep = "",
                                         min=1, max=10, value=2, step=1),
+                            downloadButton("download.map.plot", .term_switcher("export"))
                             )#end column
-                   ) #end fluid row
+                   ), #end fluid row
+                   fluidRow(column(12,
+                                   p(.term_switcher("click.on.point")),
+                                   uiOutput("map.selection.tab"),
+                                   br()
+                   )) # end colums, end fluidrow                   
           ), # end tabPanel
 
           tabPanel("Section X",  # section X ----
@@ -161,9 +176,15 @@ ui <- shinyUI(
                             sliderInput("sectionY.point.size", .term_switcher("point.size"),
                                         width="100%", sep = "",
                                         min=1, max=10, value=5, step=1),
-                            plotOutput("site.mapY")
+                            plotOutput("site.mapY"),
+                            downloadButton("download.section.y.plot", .term_switcher("export"))
                      )
-                   )#end fluidrow
+                   ), #end fluid row
+                   fluidRow(column(12,
+                                   p(.term_switcher("click.on.point")),
+                                   uiOutput("sectionY.selection.tab"),
+                                   br()
+                   )) # end colums, end fluidrow   
           ), # end tabPanel
 
           tabPanel("Section Y",  #section Y ----
@@ -184,9 +205,15 @@ ui <- shinyUI(
                             sliderInput("sectionX.point.size", .term_switcher("point.size"),
                                         width="100%", sep = "",
                                         min=1, max=10, value=5, step=1),
-                            plotOutput("site.mapX")
+                            plotOutput("site.mapX"),
+                            downloadButton("download.section.x.plot", .term_switcher("export"))
                      )
-                   ) #end fluidrow
+                   ), #end fluid row
+                   fluidRow(column(12,
+                                   p(.term_switcher("click.on.point")),
+                                   uiOutput("sectionX.selection.tab"),
+                                   br()
+                   )) # end colums, end fluidrow  
           ), # end tabPanel
 
           tabPanel(.term_switcher("tab.tables"),  #  tables ----