+++ NEWS.md
@@ -1,6 +1,27 @@
 
+# archeoViz 1.1.1
+Released: 2023-05-15
+
+* Change in the 'location' menu: add a third option to visualise spatial uncertainty in the location of the objects (as lines, plans, volumes). Consequently, chage in the pssible values for the 'location' parameter ('exact.fuzzy' is replaced by 'show.uncertainty').
+* In the 'Tables' tab, the types of spatial uncertainty can now be displayed in the tables.
+* Bugs fixed: activating the immediate display of the refitting lines (with the parameters 'map.refits', 'sectionX.refits', 'sectionX.refits') failed and has been fixed.
+
+# archeoViz 1.0.2
+Released: 2023-05-04
+
+* In the '3D plot', only the refitting lines related to the objects included in the data subset are displayed.
+* Fixing issues in the generation of the R command in the 'reproducibility' tab.
+* Add a table with details about the clicked point in the 'Map', 'Section X', and 'Section Y' tabs (using knitr::kable(), what adds a dependency to knitr).
+* Add support for html in the tables displaying details about the clicked point.
+* Add buttons to export the plots as html widgets in the 'Map', 'Section X', and 'Section Y' tabs.
+
+# archeoViz 1.0.1
+Released: 2023-03-03
+
+* fix timeline plots when the 'reverse.axis.values' parameter is used.
+
 # archeoViz 1.0
-Released: 2023-02-27
+Released: 2023-02-27 (on CRAN)
 
 * Add a 'Reproducibility' tab, generating an R command to launch archeoViz() with the current settings of the app.
 * In the '3D plot' tab, add a button to download the 3D plot widget as a .html file.